refactor: extract button components for customizable nav bottom
